// admin/src/api/blog/content-types/blog/blog.ts
// Interface automatically generated by schemas-to-ts

import { Tag } from '../../../tag/content-types/tag/tag'
import { Category } from '../../../category/content-types/category/category'
import { Media } from '../../../../common/schemas-to-ts/Media'
import { Tag_Plain } from '../../../tag/content-types/tag/tag'
import { Category_Plain } from '../../../category/content-types/category/category'
import { AdminPanelRelationPropertyModification } from '../../../../common/schemas-to-ts/AdminPanelRelationPropertyModification'

export interface Blog {
  id: number
  attributes: {
    createdAt: Date
    updatedAt: Date
    publishedAt?: Date
    title?: string
    slug?: string
    date?: Date
    author?: string
    content?: any
    tags?: { data: Tag[] }
    subtitle?: string
    category?: { data: Category }
    isFeature?: boolean
    recommendedBlogs: { data: Blog[] }
    share_image?: { data: Media }
    generated_share_image?: { data: Media }
    generated_share_text?: string
  }
}
export interface Blog_Plain {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  title?: string
  slug?: string
  date?: Date
  author?: string
  content?: any
  tags?: Tag_Plain[]
  subtitle?: string
  category?: Category_Plain
  isFeature?: boolean
  recommendedBlogs: Blog_Plain[]
  share_image?: Media
  generated_share_image?: Media
  generated_share_text?: string
}

export interface Blog_NoRelations {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  title?: string
  slug?: string
  date?: Date
  author?: string
  content?: any
  tags?: number[]
  subtitle?: string
  category?: number
  isFeature?: boolean
  recommendedBlogs: number[]
  share_image?: number
  generated_share_image?: number
  generated_share_text?: string
}

export interface Blog_AdminPanelLifeCycle {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  title?: string
  slug?: string
  date?: Date
  author?: string
  content?: any
  tags?: AdminPanelRelationPropertyModification<Tag_Plain>
  subtitle?: string
  category?: AdminPanelRelationPropertyModification<Category_Plain>
  isFeature?: boolean
  recommendedBlogs: AdminPanelRelationPropertyModification<Blog_Plain>
  share_image?: AdminPanelRelationPropertyModification<Media>
  generated_share_image?: AdminPanelRelationPropertyModification<Media>
  generated_share_text?: string
}