// admin/src/api/glossary/content-types/glossary/glossary.ts
// Interface automatically generated by schemas-to-ts

import { Paragraphs } from '../../../../components/word/interfaces/Paragraphs'
import { Paragraphs_Plain } from '../../../../components/word/interfaces/Paragraphs'
import { Paragraphs_NoRelations } from '../../../../components/word/interfaces/Paragraphs'

export enum Letter {
  A = 'A',
  B = 'B',
  C = 'C',
  D = 'D',
  E = 'E',
  F = 'F',
  G = 'G',
  H = 'H',
  I = 'I',
  J = 'J',
  K = 'K',
  L = 'L',
  M = 'M',
  N = 'N',
  O = 'O',
  P = 'P',
  Q = 'Q',
  R = 'R',
  S = 'S',
  T = 'T',
  U = 'U',
  V = 'V',
  W = 'W',
  X = 'X',
  Y = 'Y',
  Z = 'Z',
}

export interface Glossary {
  id: number
  attributes: {
    createdAt: Date
    updatedAt: Date
    publishedAt?: Date
    word: string
    description: string
    slug: string
    paragraphs: Paragraphs[]
    letter: Letter
  }
}
export interface Glossary_Plain {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  word: string
  description: string
  slug: string
  paragraphs: Paragraphs_Plain[]
  letter: Letter
}

export interface Glossary_NoRelations {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  word: string
  description: string
  slug: string
  paragraphs: Paragraphs_NoRelations[]
  letter: Letter
}

export interface Glossary_AdminPanelLifeCycle {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  word: string
  description: string
  slug: string
  paragraphs: Paragraphs_Plain[]
  letter: Letter
}